docs: add docs and extract common composables
