+++ app/src/main/java/com/github/onlynotesswent/model/flashcard/FlashcardViewModel.kt
@@ -34,8 +34,6 @@ class FlashcardViewModel(private val repository: FlashcardRepository) : ViewMode
   val noteFlashcards: StateFlow<List<Flashcard>> = _noteFlashcards.asStateFlow()
 
   /** Initializes the FlashcardViewModel and the repository. */
-  // TODO: Once we change user uid to firebase auth uid, we can call getFlashcardsFrom with
-  // the uid for the success callback
   init {
     repository.init {}
   }
