+++ app/src/main/java/com/github/onlynotesswent/model/flashcard/deck/Deck.kt
@@ -26,6 +26,9 @@ data class Deck(
     val description: String = "",
     val flashcardIds: List<String> = emptyList()
 ) {
+  /**
+   * Represents the play mode of a deck.
+   */
   enum class PlayMode {
     FLASHCARD,
     MATCH,
@@ -33,21 +36,38 @@ data class Deck(
     ALL;
 
     companion object {
+      /**
+       * Converts a string to a PlayMode.
+       *
+       * @param s The string to convert.
+       * @return The converted PlayMode.
+       */
       fun fromString(s: String?): PlayMode {
         return entries.find { it.toString() == s } ?: FLASHCARD
       }
     }
   }
 
+    /**
+     * Represents the sort mode of a deck.
+     */
   enum class SortMode {
     ALPHABETICAL,
     REVIEW,
     LEVEL;
 
+    /**
+     * Represents the order of sorting.
+     */
     enum class Order {
       HIGH_LOW,
       LOW_HIGH;
 
+        /**
+         * Returns the next order.
+         *
+         * @return The next order.
+         */
       fun next(): Order {
         return when (this) {
           HIGH_LOW -> LOW_HIGH
@@ -56,6 +76,11 @@ data class Deck(
       }
     }
 
+    /**
+     * Converts the sort mode to a readable string.
+     *
+     * @return The readable string.
+     */
     fun toReadableString(): String {
       return when (this) {
         ALPHABETICAL -> "Alphabetical"
@@ -64,6 +89,13 @@ data class Deck(
       }
     }
 
+    /**
+     * Sorts the flashcards with the given [sortOrder] by returning a sorted copy of [flashcards].
+     *
+     * @param flashcards The flashcards to sort.
+     * @param sortOrder The order of sorting.
+     * @return The sorted flashcards.
+     */
     fun sort(flashcards: List<Flashcard>, sortOrder: Order): List<Flashcard> {
       return when (this) {
         ALPHABETICAL -> flashcards.sortedBy { card -> card.front.lowercase().trim() }
@@ -74,13 +106,27 @@ data class Deck(
   }
 
   companion object {
+      // The maximum length of the description
     private const val DESCRIPTION_MAX_LEN = 200
+        // The maximum length of the title
     private const val TITLE_MAX_LEN = 75
 
+    /**
+     * Formats the description by trimming it and taking the first [DESCRIPTION_MAX_LEN] characters.
+     *
+     * @param s The description to format.
+     * @return The formatted description.
+     */
     fun formatDescription(s: String): String {
       return s.trimStart().take(DESCRIPTION_MAX_LEN)
     }
 
+    /**
+     * Formats the title by trimming it and taking the first [TITLE_MAX_LEN] characters.
+     *
+     * @param s The title to format.
+     * @return The formatted title.
+     */
     fun formatTitle(s: String): String {
       return s.trimStart().take(TITLE_MAX_LEN)
     }
